fix: graceful shutdown handler
